requirements.txt
Django>=4.2.0,<5.0.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.1
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
django-crispy-forms>=2.1
django-extensions==3.2.3
django-debug-toolbar==4.2.0
Pillow>=10.0.0
python-dotenv>=1.0.0 